Don't require metrics tasks outside development environment.

+++ lib/tasks/format_metrics.rake
@@ -2,29 +2,31 @@
 require 'nokogiri'
 require 'fileutils'
 
-namespace :metrics do
-  desc 'Generate a fancy report of all the code metrics run so far'
-  task :generate do
-    require 'rouge'
+if Rails.env.development?
 
-    # Get all the source against which we'll run any metrics
-    ruby_files = Dir['{app,config,db,lib,spec}/**/*.{rb,rake,god}']
-    ruby_files.delete('db/schema.rb')
-    ruby_files << 'Gemfile'
+  namespace :metrics do
+    desc 'Generate a fancy report of all the code metrics run so far'
+    task :generate do
+      require 'rouge'
 
-    # Run it through the pretty-printer first
-    formatter = Rouge::Formatters::HTML.new(
-      css_class: 'highlight',
-      line_numbers: true
-    )
-    lexer = Rouge::Lexers::Ruby.new
-    theme = Rouge::Themes::Base16
-    ruby_files.each do |in_file|
-      metric_filename = File.join('doc', 'metrics', in_file + '.html')
+      # Get all the source against which we'll run any metrics
+      ruby_files = Dir['{app,config,db,lib,spec}/**/*.{rb,rake,god}']
+      ruby_files.delete('db/schema.rb')
+      ruby_files << 'Gemfile'
 
-      FileUtils.mkdir_p(File.dirname(metric_filename))
-      File.open(metric_filename, 'w') do |out_file|
-        out_file.write(<<-eos
+      # Run it through the pretty-printer first
+      formatter = Rouge::Formatters::HTML.new(
+                                              css_class: 'highlight',
+                                              line_numbers: true
+                                              )
+      lexer = Rouge::Lexers::Ruby.new
+      theme = Rouge::Themes::Base16
+      ruby_files.each do |in_file|
+        metric_filename = File.join('doc', 'metrics', in_file + '.html')
+
+        FileUtils.mkdir_p(File.dirname(metric_filename))
+        File.open(metric_filename, 'w') do |out_file|
+          out_file.write(<<-eos
 <!DOCTYPE html>
 <head>
   <meta charset="utf-8">
@@ -55,140 +57,140 @@ namespace :metrics do
 <body>
 <h1>Metrics for #{in_file}</h1>
 eos
-        )
+                         )
 
-        out_file.write(formatter.format(lexer.lex(IO.read(in_file))))
-        out_file.write('</body></html>')
-      end
-    end
-
-    # Next: read in each of the metrics files and put them in a giant hash
-    # of stuff
-    metric_data = { }
-
-    if File.exist? File.join('doc', 'metrics', 'rubocop.txt')
-      current_dir = Dir.pwd
-      File.open(File.join('doc', 'metrics', 'rubocop.txt')).each_line do |l|
-        next unless l.start_with? '/'
-
-        matches = l.match(/#{current_dir}\/(.*):(\d*):\d*: (.*)/)
-        unless matches
-          puts "WARNING: Could not read Rubocop line #{l}"
-          next
+          out_file.write(formatter.format(lexer.lex(IO.read(in_file))))
+          out_file.write('</body></html>')
         end
-
-        metric_data[matches[1]] ||= { }
-        metric_data[matches[1]][Integer(matches[2])] ||= []
-        metric_data[matches[1]][Integer(matches[2])] << '<strong>RuboCop: </strong>' + matches[3]
       end
-    end
 
-    if File.exist? File.join('doc', 'metrics', 'yardstick.txt')
-      File.open(File.join('doc', 'metrics', 'yardstick.txt')).each_line do |l|
-        matches = l.match(/(.*):(\d*): (.*)/)
-        next unless matches
+      # Next: read in each of the metrics files and put them in a giant hash
+      # of stuff
+      metric_data = { }
 
-        metric_data[matches[1]] ||= { }
-        metric_data[matches[1]][Integer(matches[2])] ||= []
-        metric_data[matches[1]][Integer(matches[2])] << '<strong>Yardstick: </strong>' + matches[3]
+      if File.exist? File.join('doc', 'metrics', 'rubocop.txt')
+        current_dir = Dir.pwd
+        File.open(File.join('doc', 'metrics', 'rubocop.txt')).each_line do |l|
+          next unless l.start_with? '/'
+
+          matches = l.match(/#{current_dir}\/(.*):(\d*):\d*: (.*)/)
+          unless matches
+            puts "WARNING: Could not read Rubocop line #{l}"
+            next
+          end
+
+          metric_data[matches[1]] ||= { }
+          metric_data[matches[1]][Integer(matches[2])] ||= []
+          metric_data[matches[1]][Integer(matches[2])] << '<strong>RuboCop: </strong>' + matches[3]
+        end
       end
-    end
 
-    if File.exist? File.join('doc', 'metrics', 'excellent.html')
-      excellent_file = File.open(File.join('doc', 'metrics', 'excellent.html'))
-      excellent_doc = Nokogiri::HTML::Document.parse(excellent_file)
-      excellent_doc.css('#results .file').each do |file|
-        filename = file.at_css('dl dt').content
+      if File.exist? File.join('doc', 'metrics', 'yardstick.txt')
+        File.open(File.join('doc', 'metrics', 'yardstick.txt')).each_line do |l|
+          matches = l.match(/(.*):(\d*): (.*)/)
+          next unless matches
 
-        problems = file.css('dl dd').each do |p|
-          line = Integer(p.at_css('.lineNumber .number').content)
-          p.css('span').remove
-          problem = p.content.strip
+          metric_data[matches[1]] ||= { }
+          metric_data[matches[1]][Integer(matches[2])] ||= []
+          metric_data[matches[1]][Integer(matches[2])] << '<strong>Yardstick: </strong>' + matches[3]
+        end
+      end
+
+      if File.exist? File.join('doc', 'metrics', 'excellent.html')
+        excellent_file = File.open(File.join('doc', 'metrics', 'excellent.html'))
+        excellent_doc = Nokogiri::HTML::Document.parse(excellent_file)
+        excellent_doc.css('#results .file').each do |file|
+          filename = file.at_css('dl dt').content
+
+          problems = file.css('dl dd').each do |p|
+            line = Integer(p.at_css('.lineNumber .number').content)
+            p.css('span').remove
+            problem = p.content.strip
+
+            metric_data[filename] ||= { }
+            metric_data[filename][line] ||= []
+            metric_data[filename][line] << '<strong>Excellent: </strong>' + problem
+          end
+        end
+      end
+
+      if File.exist? File.join('doc', 'metrics', 'railsbp.html')
+        railsbp_file = File.open(File.join('doc', 'metrics', 'railsbp.html'))
+        railsbp_doc = Nokogiri::HTML::Document.parse(railsbp_file)
+        railsbp_doc.css('table.result tbody tr').each do |row|
+          filename = row.at_css('td.filename').content.strip
+          line = Integer(row.at_css('td.line').content.strip)
+          problem = row.at_css('td.message a').content.strip
 
           metric_data[filename] ||= { }
           metric_data[filename][line] ||= []
-          metric_data[filename][line] << '<strong>Excellent: </strong>' + problem
+          metric_data[filename][line] << '<strong>rails_best_practices: </strong>' + problem
         end
       end
-    end
 
-    if File.exist? File.join('doc', 'metrics', 'railsbp.html')
-      railsbp_file = File.open(File.join('doc', 'metrics', 'railsbp.html'))
-      railsbp_doc = Nokogiri::HTML::Document.parse(railsbp_file)
-      railsbp_doc.css('table.result tbody tr').each do |row|
-        filename = row.at_css('td.filename').content.strip
-        line = Integer(row.at_css('td.line').content.strip)
-        problem = row.at_css('td.message a').content.strip
+      # FIXME: Brakeman, once its Rake task has been fixed.
 
-        metric_data[filename] ||= { }
-        metric_data[filename][line] ||= []
-        metric_data[filename][line] << '<strong>rails_best_practices: </strong>' + problem
+      # Compute the percentiles for counts for files
+      file_problems = { }
+      metric_data.each do |filename, lines|
+        count = 0
+        lines.each { |line, problems| count += problems.count }
+        file_problems[filename] = count
       end
-    end
 
-    # FIXME: Brakeman, once its Rake task has been fixed.
+      file_counts = file_problems.values.sort
+      file_percentiles = [
+                          file_counts[(file_counts.count * 0.5).ceil],
+                          file_counts[(file_counts.count * 0.9).ceil]
+                         ]
 
-    # Compute the percentiles for counts for files
-    file_problems = { }
-    metric_data.each do |filename, lines|
-      count = 0
-      lines.each { |line, problems| count += problems.count }
-      file_problems[filename] = count
-    end
+      COLORS = ['#f4bf75', '#d28445', '#ac4142', '#90a959']
 
-    file_counts = file_problems.values.sort
-    file_percentiles = [
-      file_counts[(file_counts.count * 0.5).ceil],
-      file_counts[(file_counts.count * 0.9).ceil]
-    ]
+      # Finally, add the data to the HTML files
+      ruby_files.each do |in_file|
+        data = metric_data[in_file] || { }
+        filename = File.join('doc', 'metrics', in_file + '.html')
+        html = IO.read(filename)
 
-    COLORS = ['#f4bf75', '#d28445', '#ac4142', '#90a959']
+        data.each do |line, problems|
+          if problems.count <= 1
+            bg_color = COLORS[0]
+          elsif problems.count <= 3
+            bg_color = COLORS[1]
+          else
+            bg_color = COLORS[2]
+          end
 
-    # Finally, add the data to the HTML files
-    ruby_files.each do |in_file|
-      data = metric_data[in_file] || { }
-      filename = File.join('doc', 'metrics', in_file + '.html')
-      html = IO.read(filename)
+          # Sure, we can parse HTML with sed.
+          html.sub!(
+                    "<div class=\"lineno\">#{line}</div>",
+                    "<a class=\"lineno\" style=\"background-color: #{bg_color}\" href=\"#\" onclick=\"showDialog('##{line}-dialog'); return false;\">(#{problems.count}) #{line}</a>" \
+                    "<div style=\"display: none\" id=\"#{line}-dialog\" title=\"Problems on line #{line}\"><ul><li>#{problems.join('</li><li>')}</li></ul></div>"
+                    )
+        end
 
-      data.each do |line, problems|
-        if problems.count <= 1
-          bg_color = COLORS[0]
-        elsif problems.count <= 3
-          bg_color = COLORS[1]
+        # All the other lines are green (for now...)
+        html.gsub!('<div class="lineno">', "<div class=\"lineno\" style=\"background-color: #{COLORS[3]}\">")
+
+        # And colorize the header based on the file status
+        if file_problems[in_file].nil?
+          head_color = COLORS[3]
+        elsif file_problems[in_file] < file_percentiles[0]
+          head_color = COLORS[0]
+        elsif file_problems[in_file] < file_percentiles[1]
+          head_color = COLORS[1]
         else
-          bg_color = COLORS[2]
+          head_color = COLORS[2]
         end
+        html.sub!('<h1>', "<h1 style=\"color: #{head_color}\">")
+        html.sub!('</h1>', "</h1><div class=\"sep\">#{file_problems[in_file] || 0} issues…</div>")
 
-        # Sure, we can parse HTML with sed.
-        html.sub!(
-          "<div class=\"lineno\">#{line}</div>",
-          "<a class=\"lineno\" style=\"background-color: #{bg_color}\" href=\"#\" onclick=\"showDialog('##{line}-dialog'); return false;\">(#{problems.count}) #{line}</a>" \
-          "<div style=\"display: none\" id=\"#{line}-dialog\" title=\"Problems on line #{line}\"><ul><li>#{problems.join('</li><li>')}</li></ul></div>"
-        )
+        File.open(filename, 'w') { |f| f.write(html) }
       end
 
-      # All the other lines are green (for now...)
-      html.gsub!('<div class="lineno">', "<div class=\"lineno\" style=\"background-color: #{COLORS[3]}\">")
-
-      # And colorize the header based on the file status
-      if file_problems[in_file].nil?
-        head_color = COLORS[3]
-      elsif file_problems[in_file] < file_percentiles[0]
-        head_color = COLORS[0]
-      elsif file_problems[in_file] < file_percentiles[1]
-        head_color = COLORS[1]
-      else
-        head_color = COLORS[2]
-      end
-      html.sub!('<h1>', "<h1 style=\"color: #{head_color}\">")
-      html.sub!('</h1>', "</h1><div class=\"sep\">#{file_problems[in_file] || 0} issues…</div>")
-
-      File.open(filename, 'w') { |f| f.write(html) }
-    end
-
-    # Generate the file listing
-    File.open(File.join('doc', 'metrics', 'listing.html'), 'w') do |f|
-      f.write(<<-eos
+      # Generate the file listing
+      File.open(File.join('doc', 'metrics', 'listing.html'), 'w') do |f|
+        f.write(<<-eos
 <!DOCTYPE html>
 <head>
   <meta charset="utf-8">
@@ -202,28 +204,28 @@ eos
 <body>
 <ul>
 eos
-      )
+                )
 
-      ruby_files.each do |rb|
-        if file_problems[rb].nil?
-          link_color = COLORS[3]
-        elsif file_problems[rb] < file_percentiles[0]
-          link_color = COLORS[0]
-        elsif file_problems[rb] < file_percentiles[1]
-          link_color = COLORS[1]
-        else
-          link_color = COLORS[2]
+        ruby_files.each do |rb|
+          if file_problems[rb].nil?
+            link_color = COLORS[3]
+          elsif file_problems[rb] < file_percentiles[0]
+            link_color = COLORS[0]
+          elsif file_problems[rb] < file_percentiles[1]
+            link_color = COLORS[1]
+          else
+            link_color = COLORS[2]
+          end
+
+          f.write("<li><a href='#{rb}.html' target='main' style='color: #{link_color}'>#{rb}</a></li>")
         end
 
-        f.write("<li><a href='#{rb}.html' target='main' style='color: #{link_color}'>#{rb}</a></li>")
+        f.write('</ul></body></html>')
       end
 
-      f.write('</ul></body></html>')
-    end
-
-    # Generate the frameset (teehee, cringe)
-    File.open(File.join('doc', 'metrics', 'index.html'), 'w') do |f|
-      f.write(<<-eos
+      # Generate the frameset (teehee, cringe)
+      File.open(File.join('doc', 'metrics', 'index.html'), 'w') do |f|
+        f.write(<<-eos
 <!DOCTYPE html>
 <head>
   <meta charset="utf-8">
@@ -238,20 +240,22 @@ eos
 </frameset>
 </html>
 eos
-      )
+                )
+      end
+    end
+
+    desc 'Copy the metrics to the gh-pages folder'
+    task :publish, :path do |t, args|
+      path = args[:path] || File.join('..', 'gh-pages')
+      unless Dir.exist? File.join(path, '.git')
+        abort 'ERROR: You must specify the path to a checkout of the gh-pages repository (default: ../gh-pages/)'
+      end
+
+      FileUtils.rm_rf(File.join(path, 'metrics'))
+      FileUtils.cp_r(File.join('doc', 'metrics'), path)
+
+      puts 'Finished! Now commit/push the changes to the gh-pages repository.'
     end
   end
 
-  desc 'Copy the metrics to the gh-pages folder'
-  task :publish, :path do |t, args|
-    path = args[:path] || File.join('..', 'gh-pages')
-    unless Dir.exist? File.join(path, '.git')
-      abort 'ERROR: You must specify the path to a checkout of the gh-pages repository (default: ../gh-pages/)'
-    end
-
-    FileUtils.rm_rf(File.join(path, 'metrics'))
-    FileUtils.cp_r(File.join('doc', 'metrics'), path)
-
-    puts 'Finished! Now commit/push the changes to the gh-pages repository.'
-  end
 end
